Multiple changes - add leaderboard - add question search - add notifications - add feedback - add reporting - add ability to mark question as solved - add picture upload - add ability to edit questions and answers - change homepage layout - change navbar - change authentication

+++ Spring/src/main/java/sk/unibask/data/repository/EntryRepository.java
@@ -5,7 +5,7 @@ import sk.unibask.data.model.Entry;
 
 import java.util.Optional;
 
-public interface EntryRepository extends JpaRepository<Entry, Integer> {
+public interface EntryRepository extends JpaRepository<Entry, Long> {
     Optional<Entry> findById(long id);
