+++ Spring/src/main/java/sk/unibask/data/repository/VerificationCodeRepository.java
@@ -6,7 +6,7 @@ import sk.unibask.data.model.VerificationCode;
 import java.util.List;
 import java.util.Optional;
 
-public interface VerificationCodeRepository extends JpaRepository<VerificationCode, Integer> {
+public interface VerificationCodeRepository extends JpaRepository<VerificationCode, Long> {
     Optional<VerificationCode> findByEmail(String email);
 
     List<VerificationCode> findAllByEmail(String email);
